# src/doc/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API de Gerenciamento de Fazendas",
      "description": "API para gerenciar produção de leite das fazendas",
      "version": "1.0.0"
    },
    "paths": {
...
      "/production": {
        "post": {
          "summary": "Cadastrar uma nova produção de leite",
          "tags": ["Produção de Leite"],
          "security": [{ "bearerAuth": [] }],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductionInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Produção de leite cadastrada com sucesso",
              "content": {
                "application/json": {
                  "example": {
                    "message": "Produção de leite cadastrada com sucesso",
                    "production": {
                      "_id": "12345",
                      "codigo_fazenda": "67890",
                      "data": "2023-08-15T12:00:00.000Z",
                      "volume_litros": 10000,
                      "preco": 1800
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Dados inválidos",
              "content": {
                "application/json": {
                  "example": {
                    "errors": [
                      {
                        "msg": "O código da fazenda é obrigatório"
                      }
                    ]
                  }
                }
              }
            },
            "500": {
              "description": "Erro ao cadastrar a produção de leite"
            }
          }
        },
        "get": {
          "summary": "Consultar produções de leite por código de fazenda e data",
          "tags": ["Produção de Leite"],
          "security": [{ "bearerAuth": [] }],
          "parameters": [
            {
              "in": "query",
              "name": "codigo_fazenda",
              "required": true,
              "schema": {
                "type": "string"
              }
            },
            {
              "in": "query",
              "name": "data",
              "schema": {
                "type": "string",
                "format": "date"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Produções de leite encontradas",
              "content": {
                "application/json": {
                  "example": [
                    {
                      "_id": "12345",
                      "codigo_fazenda": "67890",
                      "data": "2023-08-15T12:00:00.000Z",
                      "volume_litros": 10000,
                      "preco": 1800
                    }
                  ]
                }
              }
            },
            "500": {
              "description": "Erro ao consultar as produções de leite"
            }
          }
        }
      },
      "/production/{id}": {
        "put": {
          "summary": "Atualizar dados de uma produção de leite",
          "tags": ["Produção de Leite"],
          "security": [{ "bearerAuth": [] }],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductionInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Produção de leite atualizada com sucesso",
              "content": {
                "application/json": {
                  "example": {
                    "message": "Produção de leite atualizada com sucesso",
                    "production": {
                      "_id": "12345",
                      "codigo_fazenda": "67890",
                      "data": "2023-08-15T12:00:00.000Z",
                      "volume_litros": 12000,
                      "preco": 2160
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Dados inválidos",
              "content": {
                "application/json": {
                  "example": {
                    "errors": [
                      {
                        "msg": "O volume de litros é obrigatório"
                      }
                    ]
                  }
                }
              }
            },
            "404": {
              "description": "Produção de leite não encontrada"
            },
            "500": {
              "description": "Erro ao atualizar a produção de leite"
            }
          }
        },
        "delete": {
          "summary": "Remover uma produção de leite",
          "tags": ["Produção de Leite"],
          "security": [{ "bearerAuth": [] }],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Produção de leite removida com sucesso",
              "content": {
                "application/json": {
                  "example": {
                    "message": "Produção de leite removida com sucesso"
                  }
                }
              }
            },
            "404": {
              "description": "Produção de leite não encontrada"
            },
            "500": {
              "description": "Erro ao remover a produção de leite"
            }
          }
        }
      },
      "/production/volume": {
        "get": {
          "summary": "Consulta de volume de leite entregue por fazenda, mês e ano",
          "tags": ["Produção de Leite"],
          "security": [{ "bearerAuth": [] }],
          "parameters": [
            {
              "in": "query",
              "name": "codigo_fazenda",
              "required": true,
              "schema": {
                "type": "string"
              }
            },
            {
              "in": "query",
              "name": "mes",
              "required": true,
              "schema": {
                "type": "integer"
              }
            },
            {
              "in": "query",
              "name": "ano",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Consulta de volume de leite realizada com sucesso",
              "content": {
                "application/json": {
                  "example": {
                    "dailyVolumes": {
                      "1": 5000,
                      "2": 6000,
                      "3": 7000
                    },
                    "totalVolume": 180000,
                    "averageVolume": 6000
                  }
                }
              }
            },
            "500": {
              "description": "Erro ao consultar o volume de leite"
            }
          }
        }
      },
      "/production/preco": {
        "get": {
          "summary": "Consulta de preço do leite por fazenda, mês e ano",
          "tags": ["Produção de Leite"],
          "security": [{ "bearerAuth": [] }],
          "parameters": [
            {
              "in": "query",
              "name": "codigo_fazenda",
              "required": true,
              "schema": {
                "type": "string"
              }
            },
            {
              "in": "query",
              "name": "mes",
              "required": true,
              "schema": {
                "type": "integer"
              }
            },
            {
              "in": "query",
              "name": "ano",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Consulta de preço do leite realizada com sucesso",
              "content": {
                "application/json": {
                  "example": {
                    "formattedTotalPrecoBR": "R$ 54.000,00",
                    "formattedTotalPrecoEN": "$54,000.00"
                  }
                }
              }
            },
            "404": {
              "description": "Fazenda não encontrada"
            },
            "500": {
              "description": "Erro ao consultar o preço do leite"
            }
          }
        }
      },
      "/production/preco-por-ano": {
        "get": {
          "summary": "Consulta de preço do leite por fazenda e ano",
          "tags": ["Produção de Leite"],
          "security": [{ "bearerAuth": [] }],
          "parameters": [
            {
              "in": "query",
              "name": "codigo_fazenda",
              "required": true,
              "schema": {
                "type": "string"
              }
            },
            {
              "in": "query",
              "name": "ano",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Consulta de preço do leite realizada com sucesso",
              "content": {
                "application/json": {
                  "example": {
                    "formattedMonthlyPricesBR": [
                      "R$ 6.000,00",
                      "R$ 7.200,00",
                      "R$ 7.800,00"
                    ],
                    "formattedMonthlyPricesEN": [
                      "$6,000.00",
                      "$7,200.00",
                      "$7,800.00"
                    ]
                  }
                }
              }
            },
            "404": {
              "description": "Fazenda não encontrada"
            },
            "500": {
              "description": "Erro ao consultar os preços do leite por ano"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "ProductionInput": {
        "type": "object",
        "properties": {
          "codigo_fazenda": {
            "type": "string",
            "description": "Código da fazenda associada à produção"
          },
          "data": {
            "type": "string",
            "format": "date",
            "description": "Data da produção"
          },
          "volume_litros": {
            "type": "number",
            "description": "Volume de leite em litros"
          },
          "preco": {
            "type": "number",
            "description": "Preço da produção de leite"
          }
        },
        "required": ["codigo_fazenda", "data", "volume_litros", "preco"]
      },      
...
      }
    }
  }